Updated data set processing scripts and configuration files, including modifications to MNIST, FashionMNIST, CIFAR10 and other data sets.

+++ Code/dataloader/Colorful_MNIST.py
@@ -35,7 +35,7 @@ class Colorful_MNIST(Dataset):
     def __getitem__(self, index: int) -> Tuple[Any, Any]:
         img, target = self.data[index], self.targets[index]
 
-        img = Image.fromarray(img)
+        img = Image.fromarray(np.uint8(img))
 
         if self.transform is not None:
             img = self.transform(img)
